pyproject.toml
[project]
name = "machtms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "appnope==0.1.4",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "billiard==4.2.1",
    "boto3>=1.42.39",
    "brotli==1.1.0",
    "camel-converter==4.0.1",
    "celery==5.4.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cron-descriptor==1.4.5",
    "cryptography==44.0.0",
    "cssselect2==0.8.0",
    "cython==3.0.11",
    "decorator==5.1.1",
    "dj-database-url==2.3.0",
    "django==5.1.5",
    "django-celery-beat==2.7.0",
    "django-cors-headers==4.6.0",
    "django-dirtyfields>=1.9.9",
    "django-environ==0.12.0",
    "django-extensions==3.2.3",
    "django-filter==24.3",
    "django-model-utils==5.0.0",
    "django-redis==5.4.0",
    "django-rest-knox==5.0.2",
    "django-split-settings==1.3.2",
    "django-stubs==5.1.2",
    "django-stubs-ext==5.1.2",
    "django-timezone-field==7.1",
    "djangorestframework==3.15.2",
    "djangorestframework-api-key==3.0.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-csv==3.0.2",
    "drf-spectacular==0.28.0",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "factory-boy==3.3.1",
    "faker==35.2.0",
    "fakeredis==2.27.0",
    "flake8==7.1.2",
    "fonttools==4.56.0",
    "geographiclib==2.0",
    "geopy==2.4.1",
    "google-api-python-client>=2.188.0",
    "google-auth-oauthlib>=1.2.4",
    "hiredis==3.1.0",
    "idna==3.10",
    "inflection==0.5.1",
    "ipython==8.32.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kombu==5.4.2",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "meilisearch==0.33.1",
    "model-bakery==1.20.1",
    "mypy==1.14.1",
    "mypy-extensions==1.0.0",
    "numpy==2.2.2",
    "openapi-python-client>=0.28.1",
    "openpyxl==3.1.5",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "pika>=1.3.2",
    "pillow==11.1.0",
    "prompt-toolkit==3.0.50",
    "psycopg2==2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydyf==0.11.0",
    "pyflakes==3.2.0",
    "pygments==2.19.1",
    "pylibyaml==0.1.0",
    "pymupdf>=1.26.7",
    "pyphen==0.17.2",
    "python-crontab==3.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.23.1",
    "ruff==0.9.6",
    "serpy==0.3.1",
    "setuptools==75.8.0",
    "si==0.1.1.post1",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "sysrsync==1.1.1",
    "tinycss2==1.4.0",
    "tinyhtml5==2.0.0",
    "toml==0.10.2",
    "traitlets==5.14.3",
    "types-pytz==2024.2.0.20241221",
    "types-pyyaml==6.0.12.20241230",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "unicodecsv==0.14.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uv==0.5.26",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "weasyprint",
    "webencodings==0.5.1",
    "zopfli==0.2.3.post1",
]

[tool.uv.sources]
weasyprint = { git = "https://github.com/Kozea/WeasyPrint", rev = "c74b4780149589ed801282e97a4a947362e20d6d" }

# [tool.pyright]
# extraPaths = ["./machtms"]
